Trait cc_rpc_api::permastore::PermastoreApi

pub trait PermastoreApi<Hash, BlockHash>: Sized + Send + Sync + 'static {
    fn submit_extrinsic(&self, ext: Bytes, data: Bytes) -> FutureResult<Hash>;

    fn remove_extrinsic(
        &self, 
        bytes_or_hash: Vec<ExtrinsicOrHash<Hash>>
    ) -> Result<Vec<Hash>>;

    fn remove_data(&self, chunk_root: BlockHash) -> Result<bool>;

    fn submit(&self, value: Bytes) -> Result<H256>;

    fn retrieve(&self, key: Bytes) -> Result<Option<Bytes>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>
    where
        Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
        BlockHash: Send + Sync + 'static + DeserializeOwned,
    { ... }
}

Canyon perma storage RPC API.

Required methods

fn submit_extrinsic(&self, ext: Bytes, data: Bytes) -> FutureResult<Hash>

Sepecialized submit_extrinsic for submitting the store extrinsic and transaction data.

fn remove_extrinsic(
&self,
bytes_or_hash: Vec<ExtrinsicOrHash<Hash>>
) -> Result<Vec<Hash>>

Sepecialized remove_extrinsic for removing the extrinsic and data if any.

fn remove_data(&self, chunk_root: BlockHash) -> Result<bool>

Remove the data of a transaction.

fn submit(&self, value: Bytes) -> Result<H256>

Submit the whole data of a transaction.

fn retrieve(&self, key: Bytes) -> Result<Option<Bytes>>

Fetch storage under given key.

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> where
Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
BlockHash: Send + Sync + 'static + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.